updated file structure to run base react  vite
