+++ .env.example
@@ -1,7 +1,6 @@
-# These are set for the postgres image. You shouldn't need to change them.
-POSTGRES_USER=postgres
-POSTGRES_PASSWORD=docker
-POSTGRES_DB=blueocean
-
-CLIENT_PORT=3000
-API_PORT=8000
+DEV_DATABASE_URL='LOCALDB'
+SECRET_KEY='123'
+DATABASE_URL="postgres://vidia_database_user:ETt8D6r7WL3H09LGHYJkCwATNNfEwuZd@dpg-cisqavtiuie5ebbvq74g-a.oregon-postgres.render.com/career_services_db?ssl=true"
+PORT=10000
+API='http://localhost'
+# https://production-learnlaunchnexus.onrender.com
