--- api/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "api",
-  "version": "1.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "nodemon src/server.js",
-    "start": "node src/server.js",
-    "test": "jest"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.15",
-    "supertest": "^6.3.3",
-    "vitest": "^0.29.2"
-  },
-  "dependencies": {
-    "bcrypt": "^5.1.0",
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "jest": "^29.6.2",
-    "jsonwebtoken": "^9.0.1",
-    "pg": "^8.9.0"
-  }
-}
