--- api/vite.config.js
@@ -1,8 +1,0 @@
-export default {
-  test: {
-    watch: false,
-    cache: {
-      dir: "../node_modules/.vitest",
-    },
-  },
-};
