--- client/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "client",
-  "version": "1.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite serve --host",
-    "build": "vite build",
-    "test": "vitest"
-  },
-  "devDependencies": {
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.4.3",
-    "jsdom": "^21.1.0",
-    "msw": "^1.1.0",
-    "vite": "^3.0.9",
-    "vitest": "^0.28.5"
-  },
-  "dependencies": {
-    "axios": "^0.27.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^6.15.0"
-   }
-}
