--- client/src/components/App/App.module.css
@@ -1,3 +1,0 @@
-.tasks {
-  margin-bottom: 0.5rem;
-}
