--- client/src/components/Counter/Counter.module.css
@@ -1,13 +1,0 @@
-.root {
-  border: 10px solid rgb(1, 247, 25);
-  padding: 0.5rem;
-  display: inline-block;
-}
-
-.root > * + * {
-  margin-left: 0.25rem;
-}
-
-.action {
-  width: 1.5rem;
-}
