--- client/src/components/Login/Login.jsx
@@ -1,72 +1,0 @@
-import React, { useState } from 'react'
-import { Link } from 'react-router-dom'
-
-const Login = ({ setAuth }) => {
-
-    const [inputs, setInputs] = useState({
-        email: "",
-        password: "",
-    })
-    const { email, password } = inputs
-
-    const onChange = (e) => {
-        setInputs({ ...inputs, [e.target.name]: e.target.value })
-    }
-
-    const onClick = (e) => {
-        console.log('inside onClick')
-        onSubmitForm(e)
-    }
-
-    const onSubmitForm = async (e) => {
-        e.preventDefault();
-        try {
-            const body = { email, password }
-            const response = await fetch(`${process.env.API}/authentication/login`, {
-                method: "POST",
-                headers: { "content-type": "application/json" },
-                body: JSON.stringify(body)
-
-            })
-            const parseRes = await response.json();
-            if (parseRes.token) {
-                localStorage.setItem("token", parseRes.token);
-                setAuth(true);
-            } else {
-                setAuth(false);
-            }
-        }
-        catch (error) {
-            console.error("Error message")
-        }
-    }
-
-
-    console.log(process.env.API)
-    return (
-        <>
-        <h1 className="login-page">
-            Login
-        </h1>
-
-        <form  >
-
-            <input type="text" name="email" value={email} onChange={e => onChange(e)} className="input-field">
-
-            </input>
-       
-
-        
-            <input type="password"  name="password" value={password} onChange={e => onChange(e)} className="input-field">
-             </input>
-            <button onClick={onClick} className="submit-button"> submit </button>
-
-        </form>
-        <Link to="/register">
-        Register 
-        </Link>
-        </>
-    )
-}
-
-export default Login;
