--- client/src/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-    <script type="module" src="./index.jsx"></script>
-    <link href="./app.css" rel="stylesheet" />
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
