--- client/vite.config.js
@@ -1,14 +1,0 @@
-export default {
-  root: "src",
-  server: {
-    proxy: {
-      "/api": process.env.API_URL,
-    },
-    port: process.env.PORT,
-  },
-  test: {
-    environment: "jsdom",
-    watch: false,
-  },
-};
-
