+++ package.json
@@ -1,39 +1,45 @@
 {
-  "name": "full-stack-react",
-  "version": "1.0.0",
+  "name": "learnlaunchnexus",
+  "private": true,
+  "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "deploy": "npm --prefix ./api run start",
-    "dev_deploy": "concurrently  \"npm --prefix ./api run start\"  \"npm --prefix ./client run build && npm --prefix ./client run dev\"",
-    "lint": "prettier --check .",
-    "test": "concurrently \"npm:test:*\"",
-    "ci": "npm run lint && npm run test",
-    "test:client": "npm run test --prefix=client",
-    "test:api": "npm run test --prefix=api"
-  },
-  "devDependencies": {
-    "concurrently": "^7.1.0",
-    "prettier": "^2.8.4",
-    "nodemon": "^2.0.15",
-    "supertest": "^6.3.3",
-    "vitest": "^0.29.2",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.4.3",
-    "jsdom": "^21.1.0",
-    "msw": "^1.1.0",
-    "vite": "^3.0.9"
+    "deploy": "nodemon ./api/src/server.js",
+    "dev_deploy": "concurrently \"nodemon ./api/src/server.js\" \"npm run dev\"",
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
   "dependencies": {
+    "axios": "^0.27.2",
     "bcrypt": "^5.1.0",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.1",
     "pg": "^8.9.0",
-    "axios": "^0.27.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.15.0"
+  },
+  "devDependencies": {
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@vitejs/plugin-react": "^4.0.3",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "vite": "^4.4.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "concurrently": "^7.1.0",
+    "jsdom": "^21.1.0",
+    "msw": "^1.1.0",
+    "nodemon": "^2.0.15",
+    "prettier": "^2.8.4",
+    "supertest": "^6.3.3",
+    "vitest": "^0.29.2"
   }
 }
-
