+++ src/app.css
@@ -1,0 +1,25 @@
+body {
+  margin: 0;
+  padding: 1rem;
+}
+
+.main {
+  padding: 8px;
+}
+
+.task {
+  display: flex;
+  align-items: center;
+  border: 1px solid black;
+  padding: 2px;
+  max-width: 200px;
+  background-color:black;
+}
+
+.task + .task {
+  margin-top: 4px;
+}
+
+.task > * + * {
+  margin-left: 4px;
+}
