+++ src/components/App/App.module.css
@@ -1,0 +1,3 @@
+.tasks {
+  margin-bottom: 0.5rem;
+}
