+++ src/index.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+import { createRoot } from "react-dom/client";
+import App from "./components/App/App";
+
+createRoot(document.getElementById("root")).render(<App />);
